fix: address CodeRabbit review comments
- Add contracts/out/ and contracts/cache/ to .gitignore
- Remove build artifacts from version control
- Add _verifier address validation in VoidgunPool constructor
- Extract duplicate address_to_field/u256_to_field to shared utils module
- Add edge case tests for Poseidon2 (hash4, single input, unaligned, empty)
- Add test for different pool_id values
- Document TREE_DEPTH constant linkage in merkle.nr
