+++ .gitignore
@@ -26,3 +26,7 @@ Thumbs.db
 *.so
 *.dylib
 *.dll
+
+# Foundry build output
+contracts/out/
+contracts/cache/
