--- contracts/cache/test-failures
@@ -1,1 +1,0 @@
-test_MerkleNodeSimple|test_MerkleNodeZeros|test_ZeroPermutation
