+++ contracts/test/Poseidon2.t.sol
@@ -50,4 +50,35 @@ contract Poseidon2Test is Test {
         uint256 expected = 0x28b21b8baf76eb450729177bf9f1c40afd3fabf99883153c11d8e24d2fdd9386;
         assertEq(result, expected, "Merkle node hash(123, 456) mismatch");
     }
+
+    /// @notice Test 4-input hash (used for nullifiers and key derivation)
+    function test_Hash4Inputs() public view {
+        uint256 result = hash4(2, 100, 200, 300);
+        assertTrue(result != 0, "Hash4 should produce non-zero output");
+    }
+
+    /// @notice Test that unaligned input (not 32-byte aligned) reverts
+    function test_RejectsUnalignedInput() public {
+        bytes memory unaligned = new bytes(33);
+        (bool success,) = poseidon2.staticcall(unaligned);
+        assertFalse(success, "Should reject unaligned input");
+    }
+
+    /// @notice Test single 32-byte input
+    function test_SingleInput() public view {
+        bytes memory input = abi.encodePacked(uint256(42));
+        (bool success, bytes memory output) = poseidon2.staticcall(input);
+        assertTrue(success, "Single input should succeed");
+        uint256 result = abi.decode(output, (uint256));
+        assertTrue(result != 0, "Single input hash should be non-zero");
+    }
+
+    /// @notice Test empty input handling - returns zero (identity element)
+    function test_EmptyInput() public view {
+        bytes memory empty = new bytes(0);
+        (bool success, bytes memory output) = poseidon2.staticcall(empty);
+        assertTrue(success, "Empty input should succeed");
+        uint256 result = abi.decode(output, (uint256));
+        assertEq(result, 0, "Empty input should return zero");
+    }
 }
