+++ crates/core/src/notes.rs
@@ -1,9 +1,9 @@
 use ark_bn254::Fr as Field;
-use ark_ff::PrimeField;
 use alloy_primitives::{Address, U256};
 
 use crate::keys::ReceivingKey;
 use crate::poseidon2::{hash_commitment, hash_nullifier};
+use crate::utils::{u256_to_field, address_to_field};
 
 /// A Voidgun Note (UTXO in the shielded pool)
 #[derive(Clone, Debug)]
@@ -66,12 +66,4 @@ pub struct NoteInfo {
     pub spent: bool,
 }
 
-pub fn u256_to_field(value: U256) -> Field {
-    Field::from_be_bytes_mod_order(&value.to_be_bytes::<32>())
-}
 
-pub fn address_to_field(addr: Address) -> Field {
-    let mut bytes = [0u8; 32];
-    bytes[12..32].copy_from_slice(addr.as_slice());
-    Field::from_be_bytes_mod_order(&bytes)
-}
